+++ docs/conf.py
@@ -44,7 +44,7 @@ if not settings.configured:
         'INSTALLED_APPS': INSTALLED_APPS,
         'MEDIA_ROOT': example_settings.MEDIA_ROOT,
         'MEDIA_URL': example_settings.MEDIA_URL,
-        'MIDDLEWARE_CLASSES': example_settings.MIDDLEWARE_CLASSES,
+        # 'MIDDLEWARE_CLASSES': example_settings.MIDDLEWARE_CLASSES,
         'ROOT_URLCONF': example_settings.ROOT_URLCONF,
         'SECRET_KEY': example_settings.SECRET_KEY,
         'SITE_ID': example_settings.SITE_ID,
@@ -64,6 +64,12 @@ if not settings.configured:
             'TEMPLATE_LOADERS': example_settings.TEMPLATE_LOADERS,
         })
 
+    if versions.DJANGO_GTE_1_10:
+        configure_kwargs.update({'MIDDLEWARE': example_settings.MIDDLEWARE})
+    else:
+        configure_kwargs.update(
+            {'MIDDLEWARE_CLASSES': example_settings.MIDDLEWARE_CLASSES}
+        )
 
     settings.configure(**configure_kwargs)