+++ scripts/test_django_1_9.sh
@@ -1,5 +1,0 @@
-reset
-./scripts/uninstall.sh
-#./scripts/install_django_1_9.sh
-#python examples/simple/manage.py test dummy_thumbnails --settings=settings.testing --traceback -v 3
-./runtests.py