+++ src/dummy_thumbnails/__init__.py
@@ -5,9 +5,9 @@ Dummy thumbnails for most popular Django thumbnail generators.
 """
 
 __title__ = 'dummy_thumbnails'
-__version__ = '0.2.2'
+__version__ = '0.2.3'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('default_app_config',)