+++ src/dummy_thumbnails/conf.py
@@ -10,7 +10,7 @@ from . import defaults
 
 __title__ = 'dummy_thumbnails.conf'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('get_setting',)