+++ src/dummy_thumbnails/contrib/image_importers/feed/__init__.py
@@ -7,7 +7,7 @@ feeds, that support enclosures.
 
 __title__ = 'dummy_thumbnails.contrib.image_importers.feed'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('default_app_config',)