+++ src/dummy_thumbnails/contrib/image_importers/feed/apps.py
@@ -1,22 +1,17 @@
 """
 Apps.
 """
+from django.apps import AppConfig
 
 __title__ = 'dummy_thumbnails.contrib.image_importers.feed.apps'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = ('Config',)
 
-try:
-    from django.apps import AppConfig
 
-    __all__ = ('Config',)
+class Config(AppConfig):
+    """Config."""
 
-    class Config(AppConfig):
-        """Config."""
-
-        name = 'dummy_thumbnails.contrib.image_importers.feed'
-        label = 'dummy_thumbnails_contrib_image_importers_feed'
-
-except ImportError:
-    pass
+    name = 'dummy_thumbnails.contrib.image_importers.feed'
+    label = 'dummy_thumbnails_contrib_image_importers_feed'