+++ src/dummy_thumbnails/contrib/image_importers/feed/defaults.py
@@ -5,7 +5,7 @@ Defaults module.
 """
 __title__ = 'dummy_thumbnails.contrib.image_importers.feed.defaults'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'DEFAULT_FEED_URL',