+++ src/dummy_thumbnails/contrib/image_importers/feed/management/commands/dummy_thumbnails_import_from_feed.py
@@ -15,7 +15,7 @@ from ...settings import DEFAULT_FEED_URL
 __title__ = 'dummy_thumbnails.contrib.image_importers.feed.management.' \
             'commands.dummy_thumbnails_import_from_feed'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('Command',)