+++ src/dummy_thumbnails/contrib/thumbnailers/django_imagekit/generatorlibrary.py
@@ -18,7 +18,7 @@ from ....base import get_random_image
 __title__ = 'dummy_thumbnails.contrib.thumbnailers.django_imagekit.' \
             'generatorlibrary'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'DummyThumbnail',