+++ src/dummy_thumbnails/contrib/thumbnailers/easy_thumbnails/source_generators.py
@@ -27,7 +27,7 @@ except ImportError:
 __title__ = 'dummy_thumbnails.contrib.thumbnailers.easy_thumbnails.' \
             'source_generators'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'dummy_thumbnail',