+++ src/dummy_thumbnails/defaults.py
@@ -9,7 +9,7 @@ import os
 
 __title__ = 'dummy_thumbnails.defaults'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'IMAGES_PATH',