+++ src/dummy_thumbnails/tests/base.py
@@ -15,7 +15,7 @@ import logging
 
 __title__ = 'dummy_thumbnails.tests.base'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
-__copyright__ = '2016-2017 Artur Barseghyan'
+__copyright__ = '2016-2018 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = (
     'LOG_INFO',